+++ NEWS.md
@@ -6,3 +6,7 @@
 
 * Added an efficient `write_bed` written in Rcpp and thoroughly tested against `BEDMatrix` package.
 
+# 2019-04-01 - genio 1.0.2.9000
+
+* Improved `write_bed` error message for invalid data, documentation.
+* Extended `write_bed` tests.